requirements.txt
Django==3.2.8
gunicorn==20.1.0
psycopg2-binary>=2.8
whitenoise==5.3.0
transbank-sdk==2.0.1
